Update README description and example
Changed the description to clarify differences between this and AnyType. Also changed the name of a few things in the example to more clearly differentiate between the Structs/Traits defined in the library and the ones that exists only in this example. Makes it easier to understand when looking at it for the first time without prior knowledge of what a TypeMap is.